# dashboard-backend/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "HITCON CMT 2023 Dashboard API",
    "description": "Auto generated API document",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "X-API-KEY",
      "description": "Set API token in HTTP header with key X-API-KEY"
    }
  },
  "paths": {
    "/api/v1/ctf": {
      "post": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/geocaching": {
      "post": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/dashboard": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/wall/fish": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "operationId": "post_fish",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "required": [
                "username",
                "token"
              ],
              "properties": {
                "username": {
                  "type": "string",
                  "minLength": 1
                },
                "token": {
                  "type": "string",
                  "minLength": 1
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/wall/description": {
      "post": {
        "description": "",
        "operationId": "post_description",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "required": [
                "username",
                "token",
                "description"
              ],
              "properties": {
                "username": {
                  "type": "string",
                  "minLength": 1
                },
                "token": {
                  "type": "string",
                  "minLength": 1
                },
                "description": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/wall/flag": {
      "post": {
        "description": "",
        "operationId": "post_flag",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "required": [
                "username",
                "token",
                "flagCount"
              ],
              "properties": {
                "username": {
                  "type": "string",
                  "minLength": 1
                },
                "token": {
                  "type": "string",
                  "minLength": 1
                },
                "flagCount": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/badge/{cardUid}": {
      "put": {
        "description": "",
        "operationId": "register_badge",
        "parameters": [
          {
            "name": "cardUid",
            "in": "path",
            "required": true,
            "type": "string",
            "pattern": "[0-9a-fA-F]{8}"
          },
          {
            "name": "X-API-KEY",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "HITCON"
                }
              },
              "required": [
                "username"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/api/v1/badge/popcat": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/badge/popcat/{cardUid}": {
      "put": {
        "description": "",
        "operationId": "put_badge_popcat",
        "parameters": [
          {
            "name": "cardUid",
            "in": "path",
            "required": true,
            "type": "string",
            "pattern": "[0-9a-fA-F]{8}"
          },
          {
            "name": "X-API-KEY",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "score": {
                  "type": "number",
                  "example": 10.5
                }
              },
              "required": [
                "score"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/api/v1/badge/dino": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/badge/dino/{cardUid}": {
      "put": {
        "description": "",
        "operationId": "put_badge_dino",
        "parameters": [
          {
            "name": "cardUid",
            "in": "path",
            "required": true,
            "type": "string",
            "pattern": "[0-9a-fA-F]{8}"
          },
          {
            "name": "X-API-KEY",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "score": {
                  "type": "number",
                  "example": 10.5
                }
              },
              "required": [
                "score"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/api/v1/badge/emoji": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/badge/emoji/{cardUid}": {
      "post": {
        "description": "",
        "operationId": "post_badge_emoji",
        "parameters": [
          {
            "name": "cardUid",
            "in": "path",
            "required": true,
            "type": "string",
            "pattern": "[0-9a-fA-F]{8}"
          },
          {
            "name": "X-API-KEY",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "content": {
                  "type": "string",
                  "example": "😀"
                },
                "timestamp": {
                  "type": "string",
                  "example": "2023-01-01"
                }
              },
              "required": [
                "content",
                "timestamp"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    }
  }
}